Refactor to remove previous listener and state
Problem: The 'previous' observable changing our state each time that a
message is queued, but we don't actually care about the value except
when we're publishing a message with SSB-Tribes. This means that we have
a stream open (and eventually need to implement reconnect stuff) and
relies on global mutable state that's passed around through the app.

Solution: Remove the 'previous' listener and state, and instead expect
SSB-DB to pass the relevant feed metadata to the boxer. If we're
publishing a message without the autoboxer, like the group init, we can
look up the data we need instead of trying to maintain an in-memory
database of the most recent message.